remove unit test for asyncCompose
The functionality is tested thoroughly in integration tests.
Currently, toMappable and asyncCompose split responsibility for wrapping
non-Either types before 'ingesting' them or passing them to next function. In
the future, it may be good to consolidate this responsibility in a single place.
For now, the current design makes unit tests for asyncCompose somewhat redundant.